ID-846 Fix NIH Account Link Button
