+++ src/libs/state.ts
@@ -1,6 +1,7 @@
 import { AnyPromiseFn, Atom, atom } from '@terra-ui-packages/core-utils';
 import { UserManager } from 'oidc-client-ts';
 import { AuthContextProps } from 'react-oidc-context';
+import { NihDatasetPermission } from 'src/libs/ajax/User';
 import { OidcUser } from 'src/libs/auth';
 import { getLocalStorage, getSessionStorage, staticStorageSlot } from 'src/libs/browser-storage';
 import type { WorkspaceWrapper } from 'src/libs/workspace-utils';
@@ -74,6 +75,7 @@ export type AuthState = {
   nihStatus?: {
     linkedNihUsername: string;
     linkExpireTime: number;
+    datasetPermissions: NihDatasetPermission[];
   };
   profile: TerraUserProfile;
   refreshTokenMetadata: TokenMetadata;
