+++ src/pages/Profile.js
@@ -2,13 +2,9 @@ import _ from 'lodash/fp';
 import { Fragment, useState } from 'react';
 import { div, h, h2, h3, label, p, span } from 'react-hyperscript-helpers';
 import { ClipboardButton } from 'src/components/ClipboardButton';
-import Collapse from 'src/components/Collapse';
 import { ButtonPrimary, Checkbox, IdContainer, LabeledCheckbox, Link, spinnerOverlay } from 'src/components/common';
-import { ShibbolethLink } from 'src/components/external-account-links';
 import FooterWrapper from 'src/components/FooterWrapper';
-import { icon } from 'src/components/icons';
 import { TextInput, ValidatedInput } from 'src/components/input';
-import Modal from 'src/components/Modal';
 import { PageBox, PageBoxVariants } from 'src/components/PageBox';
 import { InfoBox } from 'src/components/PopupTrigger';
 import ProfilePicture from 'src/components/ProfilePicture';
@@ -23,13 +19,13 @@ import { getConfig } from 'src/libs/config';
 import { withErrorReporting } from 'src/libs/error';
 import Events from 'src/libs/events';
 import * as Nav from 'src/libs/nav';
-import { notify } from 'src/libs/notifications';
 import allProviders from 'src/libs/providers';
-import { memoWithName, useCancellation, useOnMount, useStore } from 'src/libs/react-utils';
+import { memoWithName, useCancellation, useOnMount } from 'src/libs/react-utils';
 import { authStore, getTerraUser } from 'src/libs/state';
 import * as Style from 'src/libs/style';
 import * as Utils from 'src/libs/utils';
 import { FenceLink } from 'src/pages/profile/external-identities/FenceLink';
+import { NihLink } from 'src/pages/profile/external-identities/NihLink';
 import { SpacedSpinner } from 'src/pages/profile/SpacedSpinner';
 import validate from 'validate.js';
 
@@ -107,148 +103,6 @@ const styles = {
       marginRight: '1.2rem',
     },
   },
-};
-
-const NihLink = ({ nihToken }) => {
-  // State
-  const { nihStatus } = useStore(authStore);
-  const [isLinking, setIsLinking] = useState(false);
-  const [isConfirmUnlinkModalOpen, setIsConfirmUnlinkModalOpen] = useState(false);
-  const [isUnlinking, setIsUnlinking] = useState(false);
-
-  // Lifecycle
-  useOnMount(() => {
-    const linkNihAccount = _.flow(
-      withErrorReporting('Error linking NIH account'),
-      Utils.withBusyState(setIsLinking)
-    )(async () => {
-      const nihStatus = await Ajax().User.linkNihAccount(nihToken);
-      authStore.update(_.set(['nihStatus'], nihStatus));
-    });
-
-    if (nihToken) {
-      // Clear the query string, but use replace so the back button doesn't take the user back to the token
-      Nav.history.replace({ search: '' });
-      linkNihAccount();
-    }
-  });
-
-  // Render
-  const { linkedNihUsername, linkExpireTime, datasetPermissions } = nihStatus || {};
-
-  const [authorizedDatasets, unauthorizedDatasets] = _.flow(_.sortBy('name'), _.partition('authorized'))(datasetPermissions);
-
-  const isLinked = !!linkedNihUsername && !isLinking;
-
-  return div({ style: styles.idLink.container }, [
-    div({ style: styles.idLink.linkContentTop(isLinked) }, [
-      div({ style: { ...styles.form.title, marginBottom: 0 } }, [
-        h3({ style: { marginRight: '0.5rem', ...styles.idLink.linkName } }, ['NIH Account']),
-        h(InfoBox, [
-          'Linking with eRA Commons will allow Terra to automatically determine if you can access controlled datasets hosted in Terra (ex. TCGA) ',
-          'based on your valid dbGaP applications.',
-        ]),
-      ]),
-      Utils.cond(
-        [!nihStatus, () => h(SpacedSpinner, ['Loading NIH account status...'])],
-        [isLinking, () => h(SpacedSpinner, ['Linking NIH account...'])],
-        [!linkedNihUsername, () => div([h(ShibbolethLink, { button: true }, ['Log in to NIH'])])],
-        () =>
-          h(Fragment, [
-            div([span({ style: styles.idLink.linkDetailLabel }, ['Username:']), linkedNihUsername]),
-            div([span({ style: styles.idLink.linkDetailLabel }, ['Link Expiration:']), span([Utils.makeCompleteDate(linkExpireTime * 1000)])]),
-            div([
-              h(ShibbolethLink, ['Renew']),
-              span({ style: { margin: '0 .25rem 0' } }, [' | ']),
-              h(
-                Link,
-                {
-                  'aria-label': 'Unlink NIH account',
-                  onClick: () => setIsConfirmUnlinkModalOpen(true),
-                },
-                ['Unlink']
-              ),
-            ]),
-          ])
-      ),
-    ]),
-    isLinked &&
-      div({ style: styles.idLink.linkContentBottom }, [
-        h3({ style: { fontWeight: 500, margin: 0 } }, ['Resources']),
-        !_.isEmpty(authorizedDatasets) &&
-          h(
-            Collapse,
-            {
-              style: { marginTop: '1rem' },
-              title: 'Authorized to access',
-              titleFirst: true,
-            },
-            [
-              div({ style: { marginTop: '0.5rem' } }, [
-                _.map(({ name }) => div({ key: name, style: { lineHeight: '24px' } }, [name]), authorizedDatasets),
-              ]),
-            ]
-          ),
-        !_.isEmpty(unauthorizedDatasets) &&
-          h(
-            Collapse,
-            {
-              style: { marginTop: '1rem' },
-              title: 'Not authorized',
-              titleFirst: true,
-              afterTitle: h(InfoBox, [
-                'Your account was linked, but you are not authorized to view these controlled datasets. ',
-                'If you think you should have access, please ',
-                h(
-                  Link,
-                  {
-                    href: 'https://dbgap.ncbi.nlm.nih.gov/aa/wga.cgi?page=login',
-                    ...Utils.newTabLinkProps,
-                  },
-                  ['verify your credentials here', icon('pop-out', { size: 12, style: { marginLeft: '0.2rem', verticalAlign: 'baseline' } })]
-                ),
-                '.',
-              ]),
-            },
-            [
-              div({ style: { marginTop: '0.5rem' } }, [
-                _.map(({ name }) => div({ key: name, style: { lineHeight: '24px' } }, [name]), unauthorizedDatasets),
-              ]),
-            ]
-          ),
-      ]),
-    isConfirmUnlinkModalOpen &&
-      h(
-        Modal,
-        {
-          title: 'Confirm unlink account',
-          onDismiss: () => setIsConfirmUnlinkModalOpen(false),
-          okButton: h(
-            ButtonPrimary,
-            {
-              onClick: _.flow(
-                withErrorReporting('Error unlinking account'),
-                Utils.withBusyState(setIsUnlinking)
-              )(async () => {
-                await Ajax().User.unlinkNihAccount();
-                authStore.update(_.set('nihStatus', {}));
-                setIsConfirmUnlinkModalOpen(false);
-                notify('success', 'Successfully unlinked account', {
-                  message: 'Successfully unlinked your account from NIH.',
-                  timeout: 30000,
-                });
-              }),
-            },
-            'OK'
-          ),
-        },
-        [
-          div(['Are you sure you want to unlink from NIH?']),
-          div({ style: { marginTop: '1rem' } }, ['You will lose access to any underlying datasets. You can always re-link your account later.']),
-          isUnlinking && spinnerOverlay,
-        ]
-      ),
-  ]);
 };
 
 const PassportLinker = ({ queryParams: { state, code } = {}, provider, prettyName }) => {
